requirements.txt
requests
dingtalkchatbot
pyyaml
lxml
sqlite3
selenium
feedparser
json